+++ extraction/scriptsSQL/extractPAD_NI.sql
@@ -1,5 +1,5 @@
 SELECT
-    try_cast(replace(terseForm, ',', '.') as decimal) AS pad_ni,
+    try_cast(replace(terseForm, ',', '.') as float) AS pad_ni,
     DATEDIFF(MINUTE, :installation_date, chartTime) as temps
 FROM 
     [CISReportingDB].[dbo].[PtAssessment]
